% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agency_match.R
\name{agency_match}
\alias{agency_match}
\title{Match to Agency Index}
\usage{
agency_match(x, y, append = F, append_df)
}
\arguments{
\item{x}{a vector of potential agency names, typically the container or journal.disam column from the govscienceuseR workflow}

\item{y}{a second vector of potential agency namess, typically the author column from the govscienceuseR workflow}

\item{append}{T or F argument to whether user wants to add their own list of agencies to be indexed. Default append = F}

\item{append_df}{if append = T, a data frame with three character columns named 'State' (level of government), 'Agency' (name of agency), and 'Abbr' (abbreviation of agency) to append to the built-in index}
}
\value{
logical vector indicating an exact match to the agency index
}
\description{
Creates a logical vector as to whether an author name or publication name is an exact match to an index of government agencies
}
\examples{
dt$agency_match <- agency_match(dt$container)

}
