% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conference_match.R
\name{conference_match}
\alias{conference_match}
\title{Match to Scimago Conference Index}
\usage{
conference_match(x)
}
\arguments{
\item{x}{vector of potential conference names, typically the container or journal.disam column from the govscienceuseR workflow}
}
\value{
logical vector indicating an exact match for Scimago conference index
}
\description{
Creates a logical vector as to whether a journal name is an exact match to an index of Scimago conferences
}
\examples{
dt$conference_match <- conference_match(dt$container)

}
