% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{journal_abbr}
\alias{journal_abbr}
\title{A list of journals and their abbreviations}
\format{
A data frame with 10320 rows and 2 variables
}
\source{
\url{https://xpertscientific.com/en/journal-abbreviations/}
}
\usage{
journal_abbr
}
\description{
A dataset containing a list of journals and their abbreivations
}
\keyword{datasets}
