% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journal_disambig.R
\name{journal_disambig}
\alias{journal_disambig}
\title{Disambiguate Journal Names}
\usage{
journal_disambig(x, remove_periods = T)
}
\arguments{
\item{x}{a column containing (potential) journal names. If working through the govscienceuseR workflow, this column name is 'container'}

\item{remove_periods}{boolean for whether to remove all remaining periods (e.g., currently "^J." is replaced by "Journal of.")}
}
\value{
a vector with disambiguated journal names
}
\description{
Creates a new column for extended journal names based on known abbreviations from \href{https://xpertscientific.com/en/journal-abbreviations/}{ISO documentation from XpertScientific}
}
\examples{
dt_clean <- journal_disambig(dt$container)

}
