% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journal_match.R
\name{journal_match}
\alias{journal_match}
\title{Match to Scimago journal index}
\usage{
journal_match(x, append = F, append_df)
}
\arguments{
\item{x}{vector of potential journal names, typically the 'container' or 'journal.disam' column from the govscienceuseR workflow, but the column can be any name
#' @param append logical (T or F) argument to whether user wants to add their own list of journals to be indexed. Default append = F}

\item{append_df}{if append = T, a data frame with one character column named 'title' to append to the built-in index}
}
\value{
logical vector indicating an exact match for the journal index
}
\description{
Creates a logical vector as to whether a journal name is an exact match to an index of Scimago journals
}
\examples{
dt$journal_match <- journal_match(dt$container)


}
