% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{keras_model}
\alias{keras_model}
\title{Keras model}
\format{
A keras model object
}
\source{
\url{input URL}
}
\usage{
keras_model
}
\description{
A neural network trained on EIA data
}
\keyword{datasets}
