% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_keras_model}
\alias{load_keras_model}
\title{Load pre-trained keras model}
\usage{
load_keras_model()
}
\value{
keras model
}
\description{
Loads a keras model object pre-trained on references extracted from EISs
}
