% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepared_by.R
\name{prepared_by}
\alias{prepared_by}
\title{Reallocate 'Prepared by' Authorship}
\usage{
prepared_by(df, x, y, z)
}
\arguments{
\item{df}{a data frame or data table}

\item{x}{character, the name of the column containing (potential) journal names. If working through the govscienceuseR workflow, this column name is 'container'}

\item{y}{character, the name of the column containing author names. If working through the govscienceuseR workflow, this column name is 'author'}

\item{z}{character, the name of the column containing publisher names. If working through the govscienceuseR workflow, this column name is 'publisher'}
}
\value{
a data frame with changes made to the three input columns
}
\description{
Identified patterns commonly found in references such as "Prepared by", "Submitted to", "Report for", etc. that key into authorship and source, and reallocates the following source either to the publisher or author columns
}
\examples{
df <- prepared_by(df, 'container', 'author', 'publisher')

}
