% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scimago.c}
\alias{scimago.c}
\title{A list of conferences from the Scimago database}
\format{
A data frame with 26693 rows and 1 variable
}
\source{
\url{https://rdrr.io/github/ikashnitsky/sjrdata/man/sjr_journals.html}
}
\usage{
scimago.c
}
\description{
All conferences between 2000-2020
}
\keyword{datasets}
