% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journal_disambig.R
\name{journal_disambig}
\alias{journal_disambig}
\title{Diambiguating journal names}
\usage{
journal_disambig(df)
}
\arguments{
\item{df}{data frame with a column of potential journal names, named "container"}
}
\value{
data frame with a new column "journal.disam"
}
\description{
Creates a new column for extended journal names based on known abbreviations from LINK AND LINK
}
