% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras_classify.R
\name{keras_classify}
\alias{keras_classify}
\title{Run Keras Model on Data}
\usage{
keras_classify(df)
}
\arguments{
\item{df}{a data frame with the following column names: container, title, author, journal, doi,}
}
\value{
the input data frame with five vectors appended: a column for each classification option with probabilities of classification and a fifth column with the classification for >95\% probability (<95\% probability for any label is marked 'unsure')
}
\description{
Makes a prediction about the classification of the citation -- a journal, an agency, a conference, or not a citation. Classifications are designated based on 95\% confidence
}
\examples{
dt <- keras_classify(dt)

}
